requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
boto3==1.34.0
botocore==1.34.0
PyJWT==2.8.0
supabase==2.0.2
requests==2.31.0
tiktoken==0.5.2
python-dotenv==1.0.0
python-multipart==0.0.6
stripe==12.4.0
stripe==10.8.0
